US975260: Rename codecs-json package appropriately
https://internal.almoctane.com/ui/entity-navigation?p=131002/6001&entityType=work_item&id=975260
