+++ codec-json/src/main/java/com/github/cafapi/common/codecs/json/ObjectMapperFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafapi.common.codecs.jsonlzf;
+package com.github.cafapi.common.codecs.json;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
