+++ codec-json/src/main/resources/META-INF/services/com.github.cafapi.common.api.Decoder
@@ -1,1 +1,1 @@
-com.github.cafapi.common.codecs.jsonlzf.JsonCodec
+com.github.cafapi.common.codecs.json.JsonCodec
